Fixed data writing logic to arduino Converted float to int

+++ app/src/main/java/com/ssr/safitsafety/data/UserData.kt
@@ -1,6 +1,6 @@
 package com.ssr.safitsafety.data
 
 data class UserData(
-    val weight: Float,
+    val weight: Int,
     val age: Int,
 )
